Correção no tipo do Pixel. Após calcular o novo valor do pixel, devemos manter esse valor como Float dentro da pixelMatrix para não perder decimais quando calcular o pixel vizenho. Apenas a imagem original irá receber o pixel convertido para inteiro.

+++ src/model/Pixel.kt
@@ -1,8 +1,8 @@
 package model
 
 class Pixel (
-    var red: Int = 0,
-    var green: Int = 0,
-    var blue: Int = 0,
-    var alpha: Int = 0
+    var red: Float = 0f,
+    var green: Float = 0f,
+    var blue: Float = 0f,
+    var alpha: Float = 0f
 )
